+++ NEWS.md
@@ -1,3 +1,9 @@
+# txomics 0.1.6
+
+* `txomics` now exports `dplyr` pipe operator `%>%`
+* Some consistency changes regarding parameter names
+* `plot_heatmap()` supporting results
+
 # txomics 0.1.5
 
 * Added a `NEWS.md` file to track changes to the package.